app/requirements.txt
fdb
Flask
Flask-Bootstrap
python-dotenv